Implement taskfiles for building container images
With the taskfile common/taskfiles/podman.tasks.yml it is possible to
build, test, debug, and create a WSL distro from a Containerfile. A lot
of things are still required for this to work flawlessly but the
images/dev directory along with the Taskfile.yml show how this can be
used to build a complete WSL distro, e.g., for personal or shared
development environments. Via the taskfile remote files it should be
possible to use this mechanism from a different repository. This has not
been tested yet.