Configure rootless Podman socket
The podman socket needs to be linked upon login to /var/run/docker.sock.
For that to work without entering the sudo password I configured the
user to have passwordless sudo rights. That is obviously a security hole
for most cases but in this case it is for my personal dev container I
take that risk. If the setup-user script is reused one can include a
parameter to optionally set these rights.